removed v1 stuffs from oracle ds
