+++ cloudinit/sources/DataSourceOracle.py
@@ -155,7 +155,6 @@ class DataSourceOracle(sources.DataSource):
             ]
         )
         self._network_config_source = KlibcOracleNetworkConfigSource()
-        self._network_config_v1: dict = {"config": [], "version": 1}
         self._network_config: dict = {"ethernets": {}, "version": 2}
 
         url_params = self.get_url_params()
@@ -173,17 +172,10 @@ class DataSourceOracle(sources.DataSource):
                 KlibcOracleNetworkConfigSource(),
             )
         if not hasattr(self, "_network_config"):
-            self._network_config = {"config": [], "version": 1}
-        if not hasattr(self, "_network_config_v1"):
-            self._network_config_v1 = {"config": [], "version": 1}
+            self._network_config = {"ethernets": {}, "version": 2}
 
     def _has_network_config(self) -> bool:
-        if self._network_config:
-            if self._network_config.get("version") == 1:
-                return bool(self._network_config.get("config"))
-            elif self._network_config.get("version") == 2:
-                return bool(self._network_config.get("ethernets"))
-        return False
+        return bool(self._network_config.get("ethernets", {}))
 
     @staticmethod
     def ds_detect() -> bool:
@@ -310,9 +302,7 @@ class DataSourceOracle(sources.DataSource):
             return self._network_config
 
         set_primary = False
-        # this is v1
         if self._is_iscsi_root():
-            self._network_config_v1 = self._get_iscsi_config_v1()
             self._network_config = self._get_iscsi_config()
         if not self._has_network_config():
             LOG.warning(
@@ -398,47 +388,6 @@ class DataSourceOracle(sources.DataSource):
             else:
                 network = ipaddress.ip_network(vnic_dict["subnetCidrBlock"])
 
-            ############################# V1 HERE #############################
-
-            if is_primary:
-                if is_ipv6_only:
-                    subnets = [{"type": "dhcp6"}]
-                else:
-                    subnets = [{"type": "dhcp"}]
-            else:
-                subnets = []
-                if vnic_dict.get("privateIp"):
-                    subnets.append(
-                        {
-                            "type": "static",
-                            "address": (
-                                f"{vnic_dict['privateIp']}/"
-                                f"{network.prefixlen}"
-                            ),
-                        }
-                    )
-                if vnic_dict.get("ipv6Addresses"):
-                    subnets.append(
-                        {
-                            "type": "static",
-                            "address": (
-                                f"{vnic_dict['ipv6Addresses'][0]}/"
-                                f"{network.prefixlen}"
-                            ),
-                        }
-                    )
-            interface_config = {
-                "name": name,
-                "type": "physical",
-                "mac_address": mac_address,
-                "mtu": MTU,
-                "subnets": subnets,
-            }
-            self._network_config_v1["config"].append(interface_config)
-
-            ############################# V2 HERE #############################
-            # Why does this elif exist???
-            # Are there plans to switch to v2?
             interface_config = {
                 "mtu": MTU,
                 "match": {"macaddress": mac_address},
@@ -464,11 +413,6 @@ class DataSourceOracle(sources.DataSource):
                     )
             self._network_config["ethernets"][name] = interface_config
 
-        LOG.debug("Comparing IMDS network configs between v1 and v2")
-        compare_v1_and_v2_config_entries(
-            self._network_config_v1, self._network_config
-        )
-
 
 class DataSourceOracleNet(DataSourceOracle):
     perform_dhcp_setup = False
@@ -492,58 +436,6 @@ def convert_v1_netplan_to_v2(network_config: dict) -> dict:
         if "network" in netplan_v2_dict:
             netplan_v2_dict = netplan_v2_dict["network"]
     return netplan_v2_dict
-
-
-def move_version_to_beggining_of_nteplan_yaml_str(yaml_str: str) -> str:
-    version_line = [
-        line for line in yaml_str.split("\n") if "version:" in line
-    ][0]
-    lines = [line for line in yaml_str.split("\n") if "version:" not in line]
-    lines.insert(1, version_line)
-    return "\n".join(lines)
-
-
-def compare_v1_and_v2_config_entries(v1_config, v2_config):
-    v1_network_state = network_state.parse_net_config_data(
-        v1_config, renderer=netplan.Renderer
-    )
-    v2_network_state = network_state.parse_net_config_data(
-        v2_config, renderer=netplan.Renderer
-    )
-    v1_rendered_string = netplan.Renderer()._render_content(v1_network_state)
-    v2_rendered_string = netplan.Renderer()._render_content(v2_network_state)
-    v1_rendered_string = move_version_to_beggining_of_nteplan_yaml_str(
-        yaml_str=v1_rendered_string,
-    )
-    v2_rendered_string = move_version_to_beggining_of_nteplan_yaml_str(
-        yaml_str=v2_rendered_string,
-    )
-    LOG.debug("v1:\n%s\n\n%s", pformat(v1_config), v1_rendered_string)
-    LOG.debug("v2:\n%s\n\n%s", pformat(v2_config), v2_rendered_string)
-
-    if yaml.safe_load(v1_rendered_string) != yaml.safe_load(
-        v2_rendered_string
-    ):
-        diff = difflib.unified_diff(
-            v1_rendered_string.splitlines(),
-            v2_rendered_string.splitlines(),
-            lineterm="",
-        )
-        LOG.warning(
-            "oracle datasource v1 and v2 network config entries do not match! "
-            "diff:\n%s",
-            "\n".join(diff),
-        )
-        return False
-
-    LOG.debug("oracle datasource v1 and v2 network config entries match!")
-    return True
-
-
-def _is_ipv4_metadata_url(metadata_address: str):
-    if not metadata_address:
-        return False
-    return metadata_address.startswith(IPV4_METADATA_ROOT.split("opc")[0])
 
 
 def _read_system_uuid() -> Optional[str]:
